feat: split public and operator health exposure
