+++ openspec/changes/health-readiness-contract-prerequisite/proposal.md
@@ -1,40 +1,46 @@
 ## Why
 
-Orchard's unauthenticated readiness endpoint currently discloses build, transport, Console, runtime, licensing, check, reason, and remediation details while evaluating only an M0 subset of the readiness conditions required by `SPEC.md` section 3.1.
-Issue #115 cannot safely implement its approved public-health split until separate control-plane owners provide authoritative model, tenant, and API-key cache hydration and the existing write-path leadership surface is confirmed to provide bounded production authority.
+Orchard's unauthenticated readiness endpoint discloses build, transport, Console,
+runtime, licensing, check, reason, and remediation details while evaluating only an
+M0 subset of the readiness conditions required by `SPEC.md` §3.1. The complete
+aggregate still depends on authoritative model, tenant, and API-key cache hydration
+and a confirmed conditional leadership source, but pilot #118 must not retain the
+public disclosure while those dependencies are built.
 
 ## What Changes
 
-- Record the approved future contract for exact minimal unauthenticated `GET /health/live` and `GET /health/ready` responses.
-- Record the approved future `GET /ops/v1/health` detail route under the existing cluster-scoped Operator-or-admin authorization boundary.
-- Require one complete aggregate readiness evaluation for both public readiness and Operator health detail.
-- Record the intentional future removal of public API transport posture as a readiness gate, because `SPEC.md` section 10.7 permits `plain_http_localhost` and `SPEC.md` section 3.1 does not list transport as a readiness condition, while transport remains an Operator observation.
-- Condition that removal on confirmed fail-closed `SPEC.md` section 10.7 validation for invalid or unresolvable transport modes from any configuration source, so health never becomes the fallback validator.
-- Require the later migration to resolve the constant `controller_boot_completed` check explicitly, since it is the other current aggregate member that `SPEC.md` section 3.1 does not require.
-- Record the accepted bounded feature loss where credential-free `orchardctl status` no longer reports remote Controller version or build identity, without adding an unauthenticated route or a probe credential.
-- Identify authoritative model, tenant, and API-key cache hydration as blocking dependencies owned by separate control-plane changes.
-- Require an explicit adequacy assessment of the existing `Orchard.ControlPlane` write gate and status surface, with separate ownership only for demonstrated gaps.
-- Forbid constants, configuration flags, readiness-only caches, unrelated caches, or successful database queries from standing in for the missing authorities.
-- Define the consumer migration and validation gates that must pass when the health behavior is implemented.
-- Make no production code, test, `SPEC.md`, or runtime behavior change in this prerequisite proposal.
-- Do not add the external probe, terminal-event validator, OpenTelemetry Collector, product metrics, tracing, logs, dashboards, alerts, or Grafana resources in this change.
+- Amend the prerequisite from one blocked atomic migration to two ordered stages.
+- Stage one removes public diagnostic disclosure now: exact status-only public
+  health, authenticated Operator detail, and an explicit
+  `orchard.readiness.legacy_m0.v1` identifier for the unchanged staged predicate.
+- Stage two later replaces the legacy predicate with one complete `SPEC.md` §3.1
+  aggregate after its authoritative dependencies exist.
+- Keep the no-shim rule for the complete aggregate and prohibit claims that stage
+  one satisfies full §3.1 readiness.
+- Keep the accepted bounded `orchardctl status` feature loss and prohibit any new
+  unauthenticated version/build route or probe credential.
+- Keep external probe, telemetry, dashboards, alerts, and issue #115 observability
+  harness work out of this change.
 
-SPEC.md impact: none in this prerequisite proposal.
-`SPEC.md` remains authoritative and unchanged until the blocking authorities exist and the accepted health behavior is implemented.
+SPEC.md impact: §3.1 now defines exact public health bodies, authenticated Operator
+health detail, and the explicitly temporary legacy predicate.
 
 ## Capabilities
 
 ### New Capabilities
 
-- `controller-health-prerequisites`: Defines the approved future health boundary, its authoritative dependencies, no-shim rule, and implementation gates without changing runtime behavior.
+None.
 
 ### Modified Capabilities
 
-None.
+- `controller-health-prerequisites`: permits immediate exposure separation while
+  preserving the authoritative dependency gate and no-shim rule for the later
+  complete readiness aggregate.
 
 ## Impact
 
-- Adds a collaborator-reviewable OpenSpec contract only.
-- Requires separate control-plane change ownership for genuine cache hydration and an explicit assessment of the existing leadership surface before issue #115 implements health wiring.
-- Preserves existing Controller routes, readiness behavior, Console, CLI, packaging, documentation, tests, and `SPEC.md`.
-- Adds no dependency, vendor SDK, telemetry backend, exported identifier, or Grafana Cloud resource.
+- Public health becomes status-only before full aggregate readiness is available.
+- Operator diagnostics move to `GET /ops/v1/health` under existing authorization.
+- The incomplete predicate remains behaviorally unchanged and is labeled rather
+  than presented as full §3.1 compliance.
+- Full aggregate ownership remains with the prerequisite control-plane changes.
