+++ openspec/changes/health-readiness-contract-prerequisite/tasks.md
@@ -11,9 +11,9 @@
 
 - [ ] 2.1 Verify that each cache owner rejects constants, configuration flags, readiness-only caches, unrelated caches, and direct database queries presented as hydration evidence.
 - [ ] 2.2 Verify that the accepted leadership source derives from the same production authority used by write authorization and does not infer authority from configured role, membership, or process presence, while permitting a validated deployment mode to decide only whether the conditional leadership condition applies.
-- [ ] 2.3 Verify that issue #115 remains blocked until every dependency exposes a stable tested status interface.
-- [ ] 2.4 Re-run the issue #115 readiness-source investigation against the exact dependency heads before starting production health changes.
-- [ ] 2.5 Verify that the deferred atomic migration inventory covers Console, CLI, packaging, local-development documentation, tests, and `docs/milestones/m0-foundation.md`, and that the recorded transport-gate removal, the disposition of the constant `controller_boot_completed` check, and the credential-free `orchardctl status` identity loss are carried into that pull request.
+- [ ] 2.3 Verify that only the complete aggregate migration remains blocked until every dependency exposes a stable tested status interface; stage-one exposure separation may proceed with the labeled legacy predicate.
+- [ ] 2.4 Re-run the readiness-source investigation against the exact dependency heads before starting the stage-two aggregate migration.
+- [x] 2.5 Verify that stage one covers Console, CLI, packaging, local-development documentation, tests, and `docs/milestones/m0-foundation.md`; defer transport-gate removal and the disposition of `controller_boot_completed` to stage two.
 - [ ] 2.6 Verify that `SPEC.md` section 10.7 configuration, wrapper, and boot validation fail closed for every invalid or unresolvable public transport mode, including a `:transport_mode` application-environment value that never passes through `ORCHARD_TRANSPORT_MODE`, and assign any demonstrated gap to a separate transport change before the transport readiness gate is removed.
 
 ## 3. Validate and Hand Off
@@ -21,7 +21,7 @@
 - [ ] 3.1 Run strict OpenSpec validation after dependency identifiers are recorded.
 - [ ] 3.2 Complete an exact-diff review of this package against the exact dependency heads and resolve every blocking finding.
 - [ ] 3.3 Run the no-mistakes gate before publishing any update to this prerequisite package.
-- [ ] 3.4 Confirm the pull request changes only this OpenSpec package and excludes transient investigation, review, prompt-export, and machine-local artifacts.
-- [ ] 3.5 State in the pull request that no production behavior changes, `SPEC.md` remains unchanged, and issue #115 remains open.
-- [ ] 3.6 Hand future cache, leadership-gap, health implementation, consumer migration, and observability-harness tasks to their owning change packages rather than adding them here.
+- [ ] 3.4 Confirm the behavior-changing pull request excludes transient investigation, review, prompt-export, and machine-local artifacts.
+- [ ] 3.5 State that stage one changes exposure and `SPEC.md` without changing the readiness predicate, and that issue #115 remains open for its external acceptance harness.
+- [ ] 3.6 Hand future cache, leadership-gap, complete-aggregate, and observability-harness tasks to their owning change packages rather than adding them to stage one.
 - [ ] 3.7 After any later sync or archive, review generated main specs and remove incomplete placeholder prose such as `Purpose TBD`.
