+++ openspec/changes/pilot-health-exposure-split/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-08-03
